Prevent incorrect use of pointers
The problem with having a `pointer` property is that it's too easy to
get the pointer and have the object be destroyed immediately, before
getting a chance of even using the pointer. Therefore switched to the
Swift's array way of doing it: `withUsafePointer` methods. Also added
pointer utility methods to avoid having nested `withUsafePointer` calls.